modified calculators to use doubles, added numotu and fontsize parameters to heatmap as well as adding more sorting options. fixed bug in mothur.h setfilepositions that caused accession names to be truncated when windows line endings were used.

+++ jackknife.cpp
@@ -67,7 +67,7 @@ EstOutput Jackknife::getValues(SAbundVector* rank){
 	
 		double jack, jacklci, jackhci;
 	
-		double maxRank = (double)rank->getMaxRank();
+		int maxRank = (double)rank->getMaxRank();
 		int S = rank->getNumBins();
 
 		double N[maxOrder+1];
